fix compiler warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@26378 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/Utils/MiscUI/MyGraph.cpp
@@ -173,7 +173,7 @@ int MyGraphSeries::GetAverageDataValue() const
     if (m_dwaValues.GetSize() == 0)
         return 0;
 
-    return nTotal / m_dwaValues.GetSize();
+    return nTotal / (int)m_dwaValues.GetSize();
 }
 
 // Returns the number of data points that are not zero.
@@ -1209,8 +1209,8 @@ void MyGraph::DrawSeriesBar(CDC& dc) const
         int nMaxDataValue = max(GetMaxDataValue(), 1);
         double barTop = m_ptOrigin.y - (double)m_nYAxisHeight *
             (GetAverageDataValue() / (double)nMaxDataValue);
-        dc.MoveTo(m_ptOrigin.x, barTop);
-        VERIFY(dc.LineTo(m_ptOrigin.x + (m_nXAxisWidth - m_rcLegend.Width() - (GAP_PIXELS * 2)), barTop));
+        dc.MoveTo(m_ptOrigin.x, (int)barTop);
+        VERIFY(dc.LineTo(m_ptOrigin.x + (m_nXAxisWidth - m_rcLegend.Width() - (GAP_PIXELS * 2)), (int)barTop));
     }
 }
 
@@ -1312,8 +1312,8 @@ void MyGraph::DrawSeriesLine(CDC& dc) const
     int nMaxDataValue = max(GetMaxDataValue(), 1);
     double barTop = m_ptOrigin.y - (double)m_nYAxisHeight *
         (GetAverageDataValue() / (double)nMaxDataValue);
-    dc.MoveTo(m_ptOrigin.x, barTop);
-    VERIFY(dc.LineTo(m_ptOrigin.x + (m_nXAxisWidth - m_rcLegend.Width() - (GAP_PIXELS * 2)), barTop));
+    dc.MoveTo(m_ptOrigin.x, (int)barTop);
+    VERIFY(dc.LineTo(m_ptOrigin.x + (m_nXAxisWidth - m_rcLegend.Width() - (GAP_PIXELS * 2)), (int)barTop));
 }
 
 //
